Split transceiver mode in to USB, firmware(hackrf), and transceiver modes
