+++ firmware/common/hackrf_core.h
@@ -243,8 +243,7 @@ typedef enum {
 	TRANSCEIVER_MODE_OFF = 0,
 	TRANSCEIVER_MODE_RX = 1,
 	TRANSCEIVER_MODE_TX = 2,
-	TRANSCEIVER_MODE_SS = 3,
-	TRANSCEIVER_MODE_CPLD_UPDATE = 4
+	TRANSCEIVER_MODE_SS = 3
 } transceiver_mode_t;
 
 typedef enum {
